[OSJC-207] Supply auth token when getting pod logs
